+++ server/controllers/post.js
@@ -3,9 +3,9 @@ const prisma = new PrismaClient();
 const cloudinary = require('cloudinary').v2;
 
 cloudinary.config({
-    cloud_name: 'desjoxkzn',
-    api_key: '217771482337439',
-    api_secret: 'YIINCJ74-YMSVuC3aktAxs6zHjk'
+    cloud_name: process.env.STORAGE_NAME,
+    api_key: process.env.STORAGE_KEY,
+    api_secret: process.env.STORAGE_SECRET
 });
 
 exports.createPost = async (req, res, next) => {
@@ -80,21 +80,23 @@ exports.deletePost = async (req, res, next) => {
             id: postId
         }
     });
-
+    const { userId, role } = req.auth;
     const image = post.imgContent.split("/").pop().split(".")[0];
 
     cloudinary.uploader.destroy(`posts/${image}`,
         { resource_type: "image" },
         async function (error, result) {
-            await prisma.post.delete({
-                where: {
-                    id: postId
-                }
-            })
-                .then(res.status(200).json({
-                    message: 'Le post à été supprimé !'
-                }))
-                .catch(error => res.status(500).json({ error }));
+            if (post.authorId === userId || role === "ADMIN") {
+                await prisma.post.delete({
+                    where: {
+                        id: postId
+                    }
+                })
+                    .then(res.status(200).json({
+                        message: 'Le post à été supprimé !'
+                    }))
+                    .catch(error => res.status(500).json({ error }));
+            }
             console.log(result, error)
         });
 };
